refactor(REVIEW-REVAMP): arch compliance — relocate review-view type to domain + split over-budget files
Behavior-neutral relocations/extractions to pass check:architecture.

Layer fixes (adapters/domain must not import runtime memory/):
- Move the ReviewMessageView type family + pure morePendingReviewsLabel to
  domain/review-message-view.ts; message-action types to domain/message-actions.ts.
  memory/review-message-view.ts keeps the snapshot-driven builder + text
  fallback and re-exports the domain types for runtime importers. Channel
  renderers (slack/telegram/teams) and domain/types.ts now import the type
  from domain. cli/memory.ts drops its runtime MemoryReviewRecord import for a
  local display DTO (respects the capped cli->memory exception).

Over-budget splits (mirror existing sibling-extraction precedent):
- system-jobs.ts: review-notification emit -> memory-dreaming-job-outcome.ts
- app-memory-review.ts: single-review lookups -> app-memory-review-lookup.ts
- openapi-routes-extended.ts: memory-review route docs -> openapi-memory-review-routes.ts
- teams.ts: message-action card sinks -> teams-message-actions.ts
- telegram/channel-delivery.ts: review send -> message-action-affordances.ts
- domain/types.ts: message-action types extracted (above)
- runtime-services.ts / channel-wiring.ts: minimal comment/blank trims

architecture-exceptions.json: register the new memory lookup file for the
same Postgres-slice storage exception app-memory-review.ts already carries.
