+++ apps/core/src/app/bootstrap/runtime-services.ts
@@ -719,9 +719,7 @@ export async function startRuntimeServices(
       hashSha256Hex: (value: string) =>
         createHash('sha256').update(value, 'utf8').digest('hex'),
     });
-    // Observer digest durable send: enqueue under the shared live-send profile
-    // (idempotent on the digest's per-day key). The outbound recovery loop below
-    // does the provider send; `durablySent` settles the digest once durably sent.
+    // Observer digest durable send: enqueue under the shared live-send profile (idempotent on the per-day key); the recovery loop below sends, `durablySent` settles it.
     setObserverDigestGateway({
       enqueue: async (input) => {
         const target = resolveDurableOutboundTarget({
