+++ apps/core/src/jobs/system-jobs.ts
@@ -30,10 +30,7 @@ import {
 import { runEmbeddingBackfill } from '../memory/app-memory-backfill.js';
 import { pollAndImportProviderBatches } from '../memory/app-memory-backfill-provider-batch.js';
 import { createEmbeddingProvider } from '../memory/memory-embeddings.js';
-import type {
-  DreamingRunStatus,
-  NormalizedMemorySubject,
-} from '../memory/memory-types.js';
+import type { DreamingRunStatus } from '../memory/memory-types.js';
 import {
   DEFAULT_MEMORY_APP_ID,
   memoryAgentIdForWorkspaceFolder,
@@ -53,9 +50,8 @@ import {
 } from './system-registration-cache.js';
 import {
   appendPendingReviewContextToError,
-  buildMemoryReviewCreatedNotification,
   countPendingReviewsForNotification,
-  createdReviewIdsFromDreamSummary,
+  emitMemoryReviewCreatedContext,
   formatMemoryDreamingOutcome,
   type MemoryReviewCreatedNotification,
 } from './memory-dreaming-job-outcome.js';
@@ -491,46 +487,6 @@ async function deleteObsoleteSystemDreamingJobs(
 
 function hasUnsettledJobLease(job: Job): boolean {
   return Boolean(job.lease_run_id || job.lease_expires_at);
-}
-
-function dreamSummaryPendingCount(summary: unknown, fallback: number): number {
-  if (summary && typeof summary === 'object' && !Array.isArray(summary)) {
-    const value = (summary as Record<string, unknown>).pendingReviews;
-    if (typeof value === 'number' && Number.isFinite(value) && value >= 0) {
-      return Math.trunc(value);
-    }
-  }
-  return fallback;
-}
-
-/**
- * Turn a completed dreaming run into the typed review-created context and hand
- * it to the terminal-notification sink. No-op when the run created no reviews or
- * no sink is wired — the notification then keeps its concise summary.
- */
-async function emitMemoryReviewCreatedContext(input: {
-  dreamRun: DreamingRunStatus | undefined;
-  memory: AppMemoryService;
-  subject: NormalizedMemorySubject;
-  onNotificationContext?: (context: MemoryReviewCreatedNotification) => void;
-}): Promise<void> {
-  const sink = input.onNotificationContext;
-  if (!sink || !input.dreamRun) return;
-  const createdReviewIds = createdReviewIdsFromDreamSummary(
-    input.dreamRun.summary,
-  );
-  if (createdReviewIds.length === 0) return;
-  const pendingCount = dreamSummaryPendingCount(
-    input.dreamRun.summary,
-    createdReviewIds.length,
-  );
-  const context = await buildMemoryReviewCreatedNotification({
-    memory: input.memory,
-    subject: input.subject,
-    createdReviewIds,
-    pendingCount,
-  });
-  if (context) sink(context);
 }
 
 export async function handleSystemJob(
