+++ scripts/architecture-exceptions.json
@@ -840,6 +840,20 @@
     "maxViolations": 1
   },
   {
+    "file": "apps/core/src/memory/app-memory-review-lookup.ts",
+    "rule": "forbidden_external_import_by_layer",
+    "reason": "Single-review lookups split out of app-memory-review.ts; same Postgres-backed memory service slice, replace with repository ports in the canonical boundary cleanup.",
+    "removeByPhase": "canonical-boundary-cleanup-phase",
+    "maxViolations": 2
+  },
+  {
+    "file": "apps/core/src/memory/app-memory-review-lookup.ts",
+    "rule": "forbidden_import_by_layer",
+    "reason": "Single-review lookups split out of app-memory-review.ts; same Postgres-backed memory service slice, replace with repository ports in the canonical boundary cleanup.",
+    "removeByPhase": "canonical-boundary-cleanup-phase",
+    "maxViolations": 1
+  },
+  {
     "file": "apps/core/src/memory/maintenance-queue.ts",
     "rule": "forbidden_import_by_layer",
     "reason": "Baseline layer-boundary debt from the in-progress canonical runtime refactor; capped so new imports fail until the boundary slice removes this exception.",
